Verificador de Fraudes/requirements.txt
Flask==3.0.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pytesseract==0.3.10
Pillow==10.4.0
tldextract==5.1.2
idna==3.7
validators==0.28.3
openai==1.40.2
httpx==0.27.2
httpcore==1.0.5
opencv-python-headless==4.10.0.84
